refect: Application > Game 오케스트레이터로 이관
